Avoid to map twice the same memory to avoid clean the full cache.
 This is a temporary decision until we have a correct cleanCache
implementation. Then we will have either :
- clean cache when we map executable page. This will have the drawback
of doing useless cleaning (if the area is map only once as executable).
But the only change we can do is to remove kernel_cursor and to restart
for KERNEL_CHOOSE_START_ADDR (but we can also keep kernel_cursor but
implement memory wrapping).
- clean cache on munmap. Here again we have two choice :
 - call cleanCaches on each munmap.
 - call cleanCaches on executable munmap memory. This need to track
executable area or at least map area which can be executable.