Change compile flag from -fPIC into -mcmodel=large.
 Add of R_X86_64_GOTPCRELX/R_X86_64_REX_GOTPCRELX relocations usage
in toolchain lead to failing link of umeq.
 This is due to the fact that some code optimizations for binop
opcodes must not be perform on binop operation when pic code is in
use. Unfortunatly check in elf_x86_64_convert_load() that prevent this
optimization for pic code failed for umeq. Indeed this function use
bfd_link_pic() macro that check if we are linking shared library or
pie code. But in case of umeq we compile code in pic but we are
building a 'normal' executable. So linker apply optimizations on
these opcode (transform code + modify static relocation type) that
failed later on since we are using a custom load address above the
2Gbytes limits ....
 Anyway using fPIC flag was a bad solution, so use the correct one
which is to select large memory model.